Redesign implementation of hp
In last version of hp, the "load" return the pointer. When the caller
want to release the hazard pointer, it needs to traverse through all
hazard pointers to find the pointer.

In this version of hp, the "load" return the address where the value
is stored. Therefore, when the caller want to release the hazard pointer
, it only needs to set the data in the address to zero. However, instead
of directly using the value returned by "load", the caller needs to
dereference it first in this version.

Aside from the design of data structure, I simplify the API's name.
All of the name including "protect" is now replaced with "pr".