Add masking load function hp_pr_load_mask
In some application, the address has certain alignment pattern, so the
bits near the end of address will be zeroes. For example, if the address
align with four, the final two bits will always be zeroes. Some
implementation uses this bits as flag bits and refers to them as zeroes
when accessing the address. In order to applying to those application, I
add a new function that will clears the final bits when loading them.

Notice that instead of clear the bits which are zeroes in mask, this
function clear the bits which is ones in mask.